repositoryler klasöre ayrıldı, öğün ekleme düzeltildi, Portion adında yeni bir tablo oluşturuldu (öğün, ürün ve porsiyon bilgilerini birleşik halde tutabilmek için), artık ürün eklenirken girilen porsiyon öğün ve ürüne göre kayıt ediliyor, form kısmında ürün eklenirken daha önce eklenmiş ürünlerin girilen porsiyon değerinin bozulmaması için yardımcı bir sınıf oluşturuldu.

+++ WeightGain.UI/AdminForms/ProductsFormAdmin.cs
@@ -63,7 +63,7 @@ namespace WeightGain.UI.AdminForms
             {
                 ProductName = productName,
                 Scale = productScale,
-                Calory = Convert.ToDecimal(productCalory),
+                Calory = Convert.ToDouble(productCalory),
 
                 Category = (Category)productCategory,
             };
@@ -141,7 +141,7 @@ namespace WeightGain.UI.AdminForms
                     ProductId = (int)productId,
                     ProductName = (string)newProductName,
                     Scale = (string)newScale,
-                    Calory = (decimal)newCalory,
+                    Calory = (double)newCalory,
                     CategoryId = (int)newCategoryID,
                     Category = (Category)newCategory,
